Put depreciations computation methods in a concern
