+++ app/models/fixed_asset.rb
@@ -244,6 +244,10 @@ class FixedAsset < Ekylibre::Record::Base
     end
   end
 
+  def round(amount)
+    currency.to_currency.round amount
+  end
+
   def started_during_financial_year_exchange?
     FinancialYearExchange.opened.where('? BETWEEN started_on AND stopped_on', started_on).any?
   end
