+++ app/services/fixed_asset/transitions/sell.rb
@@ -1,6 +1,8 @@
 class FixedAsset
   module Transitions
     class Sell < Transitionable::Transition
+      include Depreciable
+
       event :sell
       from :in_use
       to :sold
@@ -35,40 +37,11 @@ class FixedAsset
 
       def can_run?
         super && resource.valid? &&
-          sold_on_during_opened_financial_year(@sold_on) &&
+          FinancialYear.on(@sold_on)&.opened? &&
           depreciations_valid?(@sold_on) &&
           resource.product &&
           resource.sale
       end
-
-      private
-
-        def split_depreciation!(depreciation, date)
-          total_amount = depreciation.amount
-          period = Accountancy::Period.new(depreciation.started_on, depreciation.stopped_on)
-          before, after = period.split date
-
-          depreciation.update! stopped_on: before.stop,
-                               amount: round(total_amount * before.days / period.days)
-
-          resource.depreciations.create! position: depreciation.position + 1,
-                                         amount: total_amount - depreciation.amount,
-                                         started_on: after.start,
-                                         stopped_on: after.stop
-        end
-
-        def round(amount)
-          resource.currency.to_currency.round amount
-        end
-
-        def depreciations_valid?(date)
-          active = resource.depreciation_on date
-          active.nil? || resource.depreciations.following(active).all? { |d| !d.has_journal_entry? }
-        end
-
-        def sold_on_during_opened_financial_year(date)
-          FinancialYear.on(date)&.opened?
-        end
     end
   end
 end
